Rename `set-timeout-test.js` to `later-test.js`.
