+++ tests/index.ts
@@ -9,9 +9,9 @@ import './defer-once-test';
 import './defer-test';
 import './events-test';
 import './join-test';
+import './later-test';
 import './multi-turn-test';
 import './queue-push-unique-test';
 import './queue-test';
 import './run-test';
-import './set-timeout-test';
 import './throttle-test';
